Refactor e2e tests, add missing assertions
